MALI-GIA driver now works, not sure if correct
The whole system now runs back and forth and fields are being updated.
I'm not sure if the updates are happening at the write time or not. In
the MALI output file, I see the same bed topo for time levels 0, 1, 2
and then it finally changes on level 3.  I'll have to think more if that
is expected behavior or not.  So there may be an unwanted lag yet in the
system, but this appears to be working well enough to start
experimenting with.